examples/pkg/exceptions/main_test.go
package exceptions

import "testing"

func TestDivide(t *testing.T) {
	type args struct {
		a float64
		b float64
	}
	tests := []struct {
		name    string
		args    args
		want    float64
		wantErr bool
	}{
		{"Does normal division", args{6, 3}, 2, false},
		{"Does long division", args{5000, 25}, 200, false},
		{"Does not divide by zero", args{5000, 0}, 0, true},
		{"Does divide zero", args{0, 19}, 0, false},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			got, err := Divide(tt.args.a, tt.args.b)
			if (err != nil) != tt.wantErr {
				t.Errorf("Divide() error = %v, wantErr %v", err, tt.wantErr)
				return
			}
			if got != tt.want {
				t.Errorf("Divide() = %v, want %v", got, tt.want)
			}
		})
	}
}
